refactor common methods
