+++ src/main/java/ChatbotRan/Deadline.java
@@ -9,7 +9,7 @@ public class Deadline extends Task {
     }
 
     public static Deadline parse(String line, int space) {
-        String[] texts = Task.parse(line, space, "/by");
+        String[] texts = Util.parse(line, space, "/by");
         if (texts != null) {
             return new Deadline(texts[0], texts[1]);
         }
