+++ src/main/java/ChatbotRan/Todo.java
@@ -1,16 +1,18 @@
 package ChatbotRan;
 
-public class Todo extends Task{
+public class Todo extends Task {
     public Todo(String contents) {
         super(contents);
     }
+
     public static Todo parse(String line, int space) {
-        String[] texts = Task.parse(line, space);
+        String[] texts = Util.parse(line, space);
         if (texts != null) {
             return new Todo(texts[0]);
         }
         return null;
     }
+
     @Override
     String getType() {
         return "T";
